requirements.txt
asgiref==3.6.0
async-timeout==4.0.2
attrs==23.1.0
autobahn==23.1.2
Automat==22.10.0
boto3==1.26.136
botocore==1.29.136
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.0
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==2.0.12
choices==0.1
constantly==15.1.0
cryptography==40.0.2
daphne==4.0.0
defusedxml==0.7.1
Deprecated==1.2.13
Django==4.2.1
django-allauth==0.54.0
django-cors-headers==4.0.0
django-debug-toolbar==4.1.0
django-model-utils==4.3.1
django-nine==0.2.7
django-oauth-toolkit==2.2.0
django-paypal==2.0
django-sendgrid-webhook==0.3.0
django-storages==1.13.2
django-svg-image-form-field==1.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
ecdsa==0.18.0
fuzzywuzzy==0.18.0
google-auth==2.21.0
google-auth-oauthlib==1.0.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
jmespath==1.0.1
jwcrypto==1.4.2
msgpack==1.0.5
oauthlib==3.2.2
packaging==23.1
Pillow==9.5.0
psycopg2-binary==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.7.0
pyOpenSSL==23.1.1
python-codicefiscale==0.8.0
python-dateutil==2.8.2
python-fsutil==0.10.0
python-gettext==5.0
python-http-client==3.3.7
python-jose==3.3.0
python-slugify==8.0.1
python3-openid==3.2.0
pytz==2022.1
redis==4.5.5
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.9
s3transfer==0.6.1
sendgrid==6.9.7
service-identity==21.1.0
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.3.0
sqlparse==0.4.2
starkbank-ecdsa==2.2.0
text-unidecode==1.3
Twisted==22.10.0
twisted-iocpsupport==1.0.3
txaio==23.1.1
typing_extensions==4.5.0
tzdata==2022.1
urllib3==1.26.9
wrapt==1.14.0
zope.interface==6.0
